requirements.txt
ollama==0.3.3
chromadb==0.5.7
langchain_community==0.3.0
langchain==0.3.0
requests==2.32.3
pypdf==5.0.0
PyPDF3==1.0.6
python-dotenv==1.0.1
streamlit==1.38.0